requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-groq
langchain-community
pydantic
docx2txt
pypdf
unstructured[local-inference]
faiss-cpu
langchain-huggingface
sentence-transformers
